requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
attrs==24.2.0
binaryornot==0.4.4
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
crashtest==0.4.1
deprecation==2.1.0
distlib==0.3.9
dulwich==0.21.7
fastapi==0.115.4
fastjsonschema==2.20.0
filelock==3.16.1
flet==0.24.1
flet-core==0.24.1
flet-runtime==0.24.1
frozenlist==1.5.0
gotrue==2.9.3
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.4
keyring==24.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
oauthlib==3.2.2
packaging==23.2
passlib==1.7.4
pexpect==4.9.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
postgrest==0.17.2
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pypng==0.20220715.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==7.4.2
RapidFuzz==3.10.1
realtime==2.0.6
repath==0.9.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
storage3==0.8.2
supabase==2.9.1
supafunc==0.6.2
text-unidecode==1.3
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.1
watchdog==4.0.2
watchfiles==0.24.0
websockets==13.1
yarl==1.17.1
